refactor: StaccatoClient를 network 패키지로 이동
